BIZ-778 move jsonld script to gatsby-ssr
